Main Program
This is a cambination of each of our mission programs put into one program. Instead of using a third color sensor, we're using buttons to control the order that we run the missions in. The middle button will call the missions one at a time in order instead of only controling one mission.